Fix newline handling for windows, refactor platform support for cursor/area
